+++ pkg/fuse/src/PoolManager.sol
@@ -3,8 +3,7 @@ pragma solidity 0.8.11;
 
 // External reference
 import { Clones } from "@openzeppelin/contracts/proxy/Clones.sol";
-import { Trust } from "@rari-capital/solmate/src/auth/Trust.sol";
-import { ERC20 } from "@rari-capital/solmate/src/erc20/ERC20.sol";
+import { ERC20 } from "@rari-capital/solmate/src/tokens/ERC20.sol";
 import { Bytes32AddressLib } from "@rari-capital/solmate/src/utils/Bytes32AddressLib.sol";
 import { PriceOracle } from "./external/PriceOracle.sol";
 
@@ -14,6 +13,7 @@ import { TargetOracle } from "./oracles/Target.sol";
 import { ZeroOracle } from "./oracles/Zero.sol";
 import { LPOracle } from "./oracles/LP.sol";
 
+import { Trust } from "@sense-finance/v1-utils/src/Trust.sol";
 import { Errors } from "@sense-finance/v1-utils/src/libs/Errors.sol";
 import { Divider } from "@sense-finance/v1-core/src/Divider.sol";
 import { BaseAdapter as Adapter } from "@sense-finance/v1-core/src/adapters/BaseAdapter.sol";
@@ -101,7 +101,7 @@ contract PoolManager is Trust {
     event PoolDeployed(string name, address comptroller, uint256 poolIndex, uint256 closeFactor, uint256 liqIncentive);
     event TargetAdded(address indexed target, address indexed cTarget);
     event SeriesAdded(address indexed zero, address indexed lpToken);
-    event SeriesQueued(address indexed adapter, uint48 indexed maturity, address indexed pool);
+    event SeriesQueued(address indexed adapter, uint256 indexed maturity, address indexed pool);
 
     constructor(
         address _fuseDirectory,
@@ -200,7 +200,7 @@ contract PoolManager is Trust {
     /// @dev called by the Periphery, which will know which pool address to set for this Series
     function queueSeries(
         address adapter,
-        uint48 maturity,
+        uint256 maturity,
         address pool
     ) external requiresTrust {
         (address zero, , , , , , , , ) = Divider(divider).series(adapter, maturity);
@@ -220,7 +220,7 @@ contract PoolManager is Trust {
 
     /// @notice open method to add queued Zeros and LPShares to Fuse pool
     /// @dev this can only be done once the yield space pool has filled its buffer and has a TWAP
-    function addSeries(address adapter, uint48 maturity) external {
+    function addSeries(address adapter, uint256 maturity) external {
         require(sStatus[adapter][maturity] == SeriesStatus.QUEUED, Errors.SeriesNotQueued);
 
         (address zero, , , , , , , , ) = Divider(divider).series(adapter, maturity);
